requirements.txt
allure-pytest>=2.14.3
selenium>=4.33.0
pytest>=8.3.5
webdriver-manager>=4.0.2
allure-python-commons>=2.14.3
requests>=2.32.4
